LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/readyState/003-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/' failed: WebSocket is closed before the connection is established.

PASS WebSockets: delete readyState
